backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
tiktoken
pandas
